requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.50.0
django-bootstrap-modal-forms==2.2.0
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-registration==3.1
django-utils-six==2.0
gunicorn==19.9.0
idna==3.3
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
